+++ client/src/App.jsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { BrowserRouter, Routes, Route } from 'react-router-dom'
 import Home from './components/Home'
-import LeetCodeWrapped from './components/LeetCodeWrapped'
+import LeetCodeWrapped from './components/Wrapped/LeetCodeWrapped'
 import Worth from './components/Worth/Worth'
 
 const App = () => {
