+++ client/src/components/Home.jsx
@@ -5,7 +5,6 @@ import { useEffect, useState } from "react";
 import { collection, getDocs, onSnapshot } from "firebase/firestore";
 import { firestore } from "../firebase";
 import { motion } from "framer-motion";
-import LeetCodeAlert from "../components/LeetCodeAlert";
 import "../App.css";
 import WorthAlert from "./Worth/WorthAlert";
 import Skeleton from "./Skeletons/Skeleton";
