+++ client/src/components/Wrapped/LeetCodeAlert.jsx
@@ -1,5 +1,5 @@
-import '../App.css';
-import '../index.css'
+import '../../App.css';
+import '../../index.css'
 import { Link } from 'react-router-dom';
 const Alert = () => {
