+++ client/src/components/Wrapped/LeetCodeWrapped.jsx
@@ -1,10 +1,10 @@
-import '../index.css'
+import '../../index.css'
 import { useState } from 'react'
 import { Link } from 'react-router-dom'
 import { ToastContainer, toast } from 'react-toastify';
 import axios from "axios";
-import SkeletonWrapped from './Skeletons/SkeletonWrapped';
-import SvgData from '../utils/svgData'
+import SkeletonWrapped from '../Skeletons/SkeletonWrapped';
+import SvgData from '../../utils/svgData'
 
 const LeetCodeWrapped = () => {
